move more ui stuff into ui folder
